+++ src/AEntity.class.cpp
@@ -1,4 +1,4 @@
-#include <AEntity.class.hpp>
+#include "../include/AEntity.class.hpp"
 
 AEntity::AEntity()
 {
@@ -29,6 +29,6 @@ void	AEntity::setYPos(int yPos)			{this->_yPos = yPos;}
 void	AEntity::setIsDead(bool dead)		{this->_isDead = dead;}
 
 eMovementDir AEntity::getDir() const		{return (this->_dir);}
-int		AEntity::getXPos() const			{return (this->_xPos);}
-int		AEntity::getYPos() const			{return (this->_yPos);}
+float		AEntity::getXPos() const			{return (this->_xPos);}
+float		AEntity::getYPos() const			{return (this->_yPos);}
 bool	AEntity::getIsDead() const			{return (this->_isDead);}
